+++ tests/Tests/Parametizer/BuiltInSubcommand/scripts/subcommands-natural-sorting.php
@@ -2,7 +2,6 @@
 
 declare(strict_types=1);
 
-use MagicPush\CliToolkit\Parametizer\Script\ScriptAbstract;
 use MagicPush\CliToolkit\Tests\Utils\TestUtils;
 
 require_once __DIR__ . '/../../../init-console.php';