+++ tests/Utils/phpunit/docker/build.sh
@@ -1,0 +1,3 @@
+#!/bin/bash
+
+docker compose build --build-arg GID=$(id --group) --build-arg UID=$(id --user)